refactor: Update package structure, add coroutines module and update database
- Moves `LifeHubDatabase` from `com.example.lifehub.data` to `com.example.lifehub.data.db`.
- Moves `Note` from `com.example.lifehub.data` to `com.example.lifehub.data.model`.
- Moves `NotesDao` from `com.example.lifehub.data` to `com.example.lifehub.data.db.dao`.
- Updates the `LifeHubDatabase` version to 2.
- Adds new `CoroutinesModule` with `IoDispatcher`, `DefaultDispatcher` and `ApplicationScope` qualifiers.
- Updates `DataModules` to reflect the changes in package structure.
- Adds `upsert` method to `NotesDao`.
- Moves Screen files to `ui/screens` package.
- Updates Note entity ID to be a String and removes auto-generate.
