--- app/src/main/java/com/example/lifehub/data/NotesDao.kt
@@ -1,8 +1,0 @@
-package com.example.lifehub.data
-
-import androidx.room.Dao
-
-@Dao
-interface NotesDao {
-
-}
